refactor: add mail provider impl
