+++ src/main/java/br/com/thiagoRDS/api_authors/providers/MailProvider/MailProvider.java
@@ -1,0 +1,7 @@
+package br.com.thiagoRDS.api_authors.providers.MailProvider;
+
+import br.com.thiagoRDS.api_authors.providers.dtos.SendMailDTO;
+
+public abstract class MailProvider {
+  public abstract void sendMail(SendMailDTO data);
+}
