+++ src/test/java/br/com/thiagoRDS/api_authors/modules/authors/controllers/ForgotMailPasswordControllerTest.java
@@ -30,7 +30,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.repositories.AuthorsReposito
 import br.com.thiagoRDS.api_authors.modules.authors.repositories.RecoveryTokensRepository;
 import br.com.thiagoRDS.api_authors.modules.utils.Convert;
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
-import br.com.thiagoRDS.api_authors.providers.MailProvider;
+import br.com.thiagoRDS.api_authors.providers.MailProvider.MailProvider;
 
 @Transactional
 @ActiveProfiles("test")
