+++ src/test/java/br/com/thiagoRDS/api_authors/modules/authors/usecases/ForgotMailPasswordTest.java
@@ -24,7 +24,7 @@ import br.com.thiagoRDS.api_authors.modules.authors.repositories.RecoveryTokensR
 import br.com.thiagoRDS.api_authors.modules.utils.MakeAuthor;
 import br.com.thiagoRDS.api_authors.modules.utils.MakeRecoveryToken;
 import br.com.thiagoRDS.api_authors.providers.CurrentContextProvider;
-import br.com.thiagoRDS.api_authors.providers.MailProvider;
+import br.com.thiagoRDS.api_authors.providers.MailProvider.MailProvider;
 
 @ExtendWith(MockitoExtension.class)
 public class ForgotMailPasswordTest {
